moved App into components and got rid of styles folder
