+++ weddingbells/src/views/index.js
@@ -1,7 +1,7 @@
 // add additional views with:
 // export { default as ${view_name} } from ${view_location};
 
-export { default as LandingPageView } from "./LandingPageView";
+export { default as LandingPageView } from "./LandingPageView/LandingPageView";
 export { default as CouplePageView } from "./CouplePageView";
 export { default as ProtectedView } from "./ProtectedView";
-export { default as WeddingCreationView } from "./WeddingCreationView";
+export { default as WeddingCreationView } from "./WeddingCreationView/WeddingCreationView";
