--- weddingbells/src/views/landingpageview.scss
@@ -1,22 +1,0 @@
-.landingpage_wrapper .top {
-	// sizing & placement
-	height: 100vh;
-	width: 100%;
-
-	// background
-	background-image: url("https://images.pexels.com/photos/1683989/pexels-photo-1683989.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=750&w=1260");
-	background-repeat: no-repeat;
-	background-size: cover;
-
-	// display
-	display: flex;
-	flex-direction: row;
-	flex-wrap: wrap;
-	justify-content: space-between;
-	align-items: center;
-}
-
-.landingpage_wrapper .top .calltoaction {
-	// sizing & placement
-	margin: auto auto;
-}
